Moved utility function to utils
